Restructure ui src files and correct respective imports.
