+++ src/ui/components/panels/ConsolePanel.vue
