+++ src/ui/components/panels/EditorPanel.vue
