+++ src/ui/components/panels/PreviewPanel.vue
