+++ src/ui/composables/useTextures.ts
@@ -1,5 +1,5 @@
 import { ref, computed } from 'vue'
-import { DEFAULT_TEXTURES } from '../webgpu/textures'
+import { DEFAULT_TEXTURES } from '../../webgpu/textures'
 
 /**
  * Vue composable for managing texture inputs in the shader editor.
